+++ api-scanner/src/test/java/net/corda/plugins/GradleProject.java
@@ -5,10 +5,6 @@ import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.BuildTask;
 import org.gradle.testkit.runner.GradleRunner;
 import org.gradle.testkit.runner.TaskOutcome;
-import org.junit.rules.TemporaryFolder;
-import org.junit.rules.TestRule;
-import org.junit.runner.Description;
-import org.junit.runners.model.Statement;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
@@ -23,16 +19,16 @@ import static java.util.Collections.emptyList;
 import static net.corda.plugins.CopyUtils.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.gradle.testkit.runner.TaskOutcome.*;
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * JUnit rule to execute the scanApi Gradle task. This rule should be chained with TemporaryFolder.
+ * Executes the API Scanner plugin in a test Gradle project.
  */
 @SuppressWarnings("WeakerAccess")
-public class GradleProject implements TestRule {
+public class GradleProject {
     private static final String TEST_GRADLE_USER_HOME = System.getProperty("test.gradle.user.home", "");
 
-    private final TemporaryFolder projectDir;
+    private final Path projectDir;
     private final String name;
     private String taskName = "scanApi";
     private TaskOutcome expectedOutcome = SUCCESS;
@@ -41,7 +37,7 @@ public class GradleProject implements TestRule {
     private String output;
     private Path api;
 
-    public GradleProject(TemporaryFolder projectDir, String name) {
+    public GradleProject(Path projectDir, String name) {
         this.projectDir = projectDir;
         this.name = name;
         this.output = "";
@@ -87,34 +83,29 @@ public class GradleProject implements TestRule {
         }
     }
 
-    @Override
-    public Statement apply(Statement base, Description description) {
-        return new Statement() {
-            @Override
-            public void evaluate() throws Throwable {
-                installResource(projectDir, name + "/build.gradle");
-                installResource(projectDir, "repositories.gradle");
-                installResource(projectDir, "settings.gradle");
-                installResource(projectDir, "gradle.properties");
+    public GradleProject build() throws IOException {
+        installResource(projectDir, name + "/build.gradle");
+        installResource(projectDir, "repositories.gradle");
+        installResource(projectDir, "settings.gradle");
+        installResource(projectDir, "gradle.properties");
 
-                result = GradleRunner.create()
-                    .withProjectDir(projectDir.getRoot())
-                    .withArguments(getGradleArgsForTasks(taskName))
-                    .withPluginClasspath()
-                    .build();
-                output = result.getOutput();
-                System.out.println(output);
+        result = GradleRunner.create()
+            .withProjectDir(projectDir.toFile())
+            .withArguments(getGradleArgsForTasks(taskName))
+            .withPluginClasspath()
+            .build();
+        output = result.getOutput();
+        System.out.println(output);
 
-                assertEquals(expectedOutcome, getOutcomeOf(taskName));
+        assertEquals(expectedOutcome, getOutcomeOf(taskName));
 
-                api = pathOf(projectDir, "build", "api", name + ".txt");
-                validate(assertThat(api));
-            }
-        };
+        api = pathOf(projectDir, "build", "api", name + ".txt");
+        validate(assertThat(api));
+        return this;
     }
 
-    public static Path pathOf(@Nonnull TemporaryFolder folder, String... elements) {
-        return Paths.get(folder.getRoot().getAbsolutePath(), elements);
+    public static Path pathOf(@Nonnull Path folder, String... elements) {
+        return Paths.get(folder.toAbsolutePath().toString(), elements);
     }
 
     public static List<String> getGradleArgsForTasks(@Nonnull String... taskNames) {